requirements.txt
torch
torchvision
transformers
sentence-transformers
pillow
matplotlib